parking-cv-service/requirements.txt
# Core dependencies
opencv-python==4.8.1.78
numpy==1.24.3
torch==2.0.1
torchvision==0.15.2
ultralytics==8.0.200

# API and HTTP
requests==2.31.0
python-dotenv==1.0.0

# OCR
easyocr==1.7.0
pillow==10.0.1

# Utilities
pyyaml==6.0.1
